+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Upgraded
-- "calypsonet-terminal-calypso-crypto-legacysam-java-api" to version `0.2.0`.
+- "Calypsonet Terminal Calypso Crypto Legacy SAM API" to version `0.2.0`.
+- "Google Gson Library" (com.google.code.gson) to version `2.10.1`.
  
 ## [0.2.0] - 2022-12-13
 ### Added